added express sessions
express-sessions keep track of session info easier so this is more efficient when going between pages than what I had originally. checkAuth is a middleware function that will check if the user session exists (it only is created after the user is authenticated). checkSession will be expanded upon to do more than just state that the user is authenticated on each page. the logout process is a lot more efficient by simply destroying the session and redirecting the user back to the home page